@@ -1,13 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Google.Cloud.Functions.Hosting" Version="1.0.0" />
-    <PackageReference Include="Google.Cloud.Storage.V1" Version="4.1.0" />
-    <PackageReference Include="GrapeCity.DioDocs.Pdf.ja" Version="5.2.0.800" />
+    <PackageReference Include="DioDocs.Pdf.ja" Version="7.0.3" />
+    <PackageReference Include="Google.Cloud.Functions.Hosting" Version="2.1.0" />
+    <PackageReference Include="Google.Cloud.Storage.V1" Version="4.10.0" />
     <None Include="appsettings*.json" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
 </Project>
